+++ tests/quantum-cloud.test.ts
@@ -36,6 +36,8 @@ function makeState(): SimState {
     algoIdx: 0,
     songIdx: 0,
     algoStep: 0,
+    algoMode: 'single',
+    algoTimer: 0,
     frame: 0,
     elapsed: 0,
   };